Moving some stuff around and filling in expectations
while getting familiar
