+++ .ruby-version
@@ -1,1 +1,1 @@
-2.5.1
+2.6.1
