+++ .simplecov
@@ -1,0 +1,5 @@
+if ENV['COVERAGE']
+  SimpleCov.start do
+    add_filter 'test/'
+  end
+end
