+++ Gemfile
@@ -1,13 +1,17 @@
 source 'https://rubygems.org'
 
-ruby '2.5.1'
+ruby File.read('.ruby-version').strip if ENV.key?('DYNO')
 
-gem 'sinatra'
-gem 'redis'
 gem 'google-api-client', '~> 0.11'
-gem 'net-ssh'
-gem 'mocha'
-gem 'rake'
-gem 'test-unit'
-gem 'rack-test'
 gem 'libhoney'
+gem 'minitest', group: :test
+gem 'minitest-hooks', group: :test
+gem 'mocha', group: :test
+gem 'net-ssh'
+gem 'puma'
+gem 'rack-test', group: :test
+gem 'rake'
+gem 'redis'
+gem 'simplecov', group: :test
+gem 'sinatra'
+gem 'test-unit', group: :test
