+++ test/test_instance_checker.rb
@@ -1,33 +1,32 @@
-# This has to come before the requires because the thread runs at the very top level
-# which we might want to change at some point
-ENV['UNIT_TESTS'] = 'true'
+require 'simplecov'
 
-require 'test/unit'
 require 'json'
-require 'mocha/test_unit'
-require_relative '../instance_checker'
 
-class InstanceChecker
-  public :get_num_warmed_instances, :create_instance, :delete_instance, :random_zone, :pools, :redis, :compute
-end
+require 'minitest/autorun'
+require 'minitest/hooks'
+require 'mocha/minitest'
 
-class InstanceCheckerTest < Test::Unit::TestCase
+libdir = File.expand_path('../lib', __dir__)
+$LOAD_PATH.unshift(libdir) unless $LOAD_PATH.include?(libdir)
 
-  # TODO: if we want to test the redis methods, we should set up a repeatable way of doing that
+require 'warmer'
+
+class InstanceCheckerTest < Minitest::Test
+  include Minitest::Hooks
+
+  def after_all
+    super
+    clean_up_test_instances
+  end
 
   def test_create_and_verify_instance
-    instance_checker = InstanceChecker.new
-    pool = instance_checker.pools.first
-    zone = instance_checker.random_zone
+    instance_checker = Warmer::InstanceChecker.new
+    pool = Warmer.pools.first
+    zone = instance_checker.send(:random_zone)
     new_instance_info = instance_checker.create_instance(pool, zone, {"warmth": "test"})
 
-    begin
-      new_name = new_instance_info[:name]
-      assert_not_nil(new_name)
-    rescue NoMethodError => e
-      puts "Instance creation probably timed out, deleting test instance..."
-      clean_up_test_instances
-    end
+    new_name = new_instance_info[:name]
+    assert_not_nil(new_name)
 
     num = instance_checker.get_num_warmed_instances("labels.warmth:test")
 
@@ -36,55 +35,49 @@ class InstanceCheckerTest < Test::Unit::TestCase
     new_num = instance_checker.get_num_warmed_instances("labels.warmth:test")
 
     assert_equal(1, num)
-    begin
-      assert_equal(0, new_num)
-    rescue
-      puts "Assertion(s) failed which likely means that an instance didn't get deleted in time, let's clean them up"
-      clean_up_test_instances
-      # Honestly if the deletion timed out we probably don't care
-    end
+    assert_equal(0, new_num)
   end
 
   def test_clean_up_orphans
-    instance_checker = InstanceChecker.new
+    instance_checker = Warmer::InstanceChecker.new
     orphan = {
       name: "test-orphan-name",
       zone: "us-central1-b"
     }
-    instance_checker.redis.rpush('orphaned-test', JSON.dump(orphan))
-    instance_checker.redis.rpush('orphaned-test', JSON.dump(orphan))
+    Warmer.redis.rpush('orphaned-test', JSON.dump(orphan))
+    Warmer.redis.rpush('orphaned-test', JSON.dump(orphan))
     instance_checker.stubs(:delete_instance).returns(true)
 
     instance_checker.clean_up_orphans('orphaned-test')
-    assert_equal(0, instance_checker.redis.llen('orphaned-test'))
+    assert_equal(0, Warmer.redis.llen('orphaned-test'))
 
-    instance_checker.redis.del('orphaned-test')
+    Warmer.redis.del('orphaned-test')
   end
 
   # TODO test mocking out timeouts/errors from GCP
 
   def test_pool_refresh
-    instance_checker = InstanceChecker.new
-    pools = instance_checker.pools
+    instance_checker = Warmer::InstanceChecker.new
+    pools = Warmer.pools
     old_size = pools.size
 
-    instance_checker.redis.hset('poolconfigs', "foobar", 1)
+    Warmer.redis.hset('poolconfigs', "foobar", 1)
     sleep 70
-    pools = instance_checker.pools
+    pools = Warmer.pools
     assert_equal(old_size + 1, pools.size)
 
-    instance_checker.redis.hdel('poolconfigs', "foobar")
+    Warmer.redis.hdel('poolconfigs', "foobar")
   end
 
-  def clean_up_test_instances
-    instance_checker = InstanceChecker.new
+  private def clean_up_test_instances
+    instance_checker = Warmer::InstanceChecker.new
     zones = [
           "#{ENV['GOOGLE_CLOUD_REGION']}-a",
           "#{ENV['GOOGLE_CLOUD_REGION']}-b",
           "#{ENV['GOOGLE_CLOUD_REGION']}-c",
           "#{ENV['GOOGLE_CLOUD_REGION']}-f"]
     zones.each do |zone|
-      instances = instance_checker.compute.list_instances(
+      instances = Warmer.compute.list_instances(
           ENV['GOOGLE_CLOUD_PROJECT'], zone,
           filter: "labels.warmth:test")
       unless instances.items.nil?
@@ -94,8 +87,4 @@ class InstanceCheckerTest < Test::Unit::TestCase
       end
     end
   end
-
 end
-
-# unset this at the end
-ENV['UNIT_TESTS'] = nil
